Fix thread view
